+++ jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  setupFilesAfterEnv: ['./jest.setup.js'],
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+  },
+  collectCoverage: true,
+  collectCoverageFrom: ['<rootDir>/src/**/*.js']
+};