+++ src/server/repositories/index.js
@@ -1,0 +1,7 @@
+import exercisesResults from './exercises-results';
+import healthcheck from './healthcheck';
+
+export default {
+  exercisesResults,
+  healthcheck
+};